+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## _(in-progress)_
 * Change: Note compatibility through WP 5.7+
 * Change: Update copyright date (2021)
+* New: Add unit tests to verify search is removed from admin bar
 
 ## 1.8.1 _(2020-09-07)_
 * Change: Restructure unit test file structure